+++ src/main/java/boof/parser/Parser.java
@@ -1,4 +1,4 @@
-package parser;
+package boof.parser;
 public class Parser {
   public static String getCommand(String userInput) {
     String[] parts = userInput.trim().split(" ", 2);
