+++ src/main/java/boof/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package boof.task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
