+++ src/main/java/boof/task/Task.java
@@ -1,4 +1,5 @@
-package task;
+package boof.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
@@ -6,6 +7,14 @@ public class Task {
     public Task(String description) {
         this.description = description;
         this.isDone = false;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public boolean isDone() {
+        return isDone;
     }
 
     public String getStatusIcon() {
