+++ src/main/java/boof/task/TaskList.java
@@ -1,4 +1,4 @@
-package task;
+package boof.task;
 import java.util.ArrayList;
 
 public class TaskList {
