+++ src/main/java/boof/task/Todo.java
@@ -1,4 +1,4 @@
-package task;
+package boof.task;
 public class Todo extends Task {
   
   public Todo(String description) {
